# [PERF][Qwen3-next] Speedup `chunk_gated_delta_rule_fwd_h` (part of GDN attn)