## Purpose
 `chunk_gated_delta_rule_fwd_h` is a part of GDN attn. Triton implementation of  `chunk_gated_delta_rule_fwd_h` has too small parallelization for Qwen3-next case - grid size is 32. 
Add more parallelization through seqlen dimation. 

## Test Plan
To estimate performance improvement I used the following script
Compare `chunknum_T=1` what corresponds to previous version vs dynamic choose of  `chunknum_T`

<details><summary>Script</summary>

```
#!/usr/bin/env python3
"""Simple benchmark comparing chunknum_T=1 vs chunknum_T=None with various sequence lengths"""

import torch
import time
import numpy as np
from vllm.model_executor.layers.fla.ops.chunk_delta_h import chunk_gated_delta_rule_fwd_h

def create_test_inputs(T, device='cuda'):
    """Create test inputs with variable sequence length"""
    B, Hg, K = 1, 4, 128
    H, V = 8, 128
    
    k = torch.randn(B, T, Hg, K, device=device, dtype=torch.float32)
    u = torch.randn(B, T, H, V, device=device, dtype=torch.float32)
    w = torch.randn(B, T, H, K, device=device, dtype=torch.float32)
    g = torch.randn(B, T, H, device=device, dtype=torch.float32)
    initial_state = torch.randn(B, H, K, V, device=device, dtype=torch.float32)
    cu_seqlens = torch.tensor([0, T], device=device, dtype=torch.int32)
    
    return k, u, w, g, initial_state, cu_seqlens

def benchmark_kernel(k, u, w, g, initial_state, cu_seqlens, chunknum_T, num_warmup=5, num_runs=20):
    """Benchmark the kernel with specified chunknum_T"""
    
    # Warmup
    for _ in range(num_warmup):
        _ = chunk_gated_delta_rule_fwd_h(
            k=k, w=w, u=u, g=g,
            initial_state=initial_state,
            output_final_state=True,
            cu_seqlens=cu_seqlens,
            chunknum_T=chunknum_T
        )
    
    # Synchronize before timing
    torch.cuda.synchronize()
    
    # Time the runs
    times = []
    for _ in range(num_runs):
        start = time.time()
        h, v_new, final_state = chunk_gated_delta_rule_fwd_h(
            k=k, w=w, u=u, g=g,
            initial_state=initial_state,
            output_final_state=True,
            cu_seqlens=cu_seqlens,
            chunknum_T=chunknum_T
        )
        torch.cuda.synchronize()
        times.append((time.time() - start) * 1000)  # Convert to ms
    
    avg_time = np.mean(times)
    std_time = np.std(times)
    
    return avg_time, std_time

def main():
    # Set device
    device = torch.cuda.current_device()
    device_props = torch.cuda.get_device_properties(device)
    print(f"Running on: {device_props.name} (SM count: {device_props.multi_processor_count})")
    print("=" * 80)
    
    # Test different sequence lengths
    sequence_lengths = [1024, 2048, 4096, 8192, 16384]
    
    print("Testing different sequence lengths...")
    print(f"{'Seq Length':<12} {'chunknum_T=1':<25} {'chunknum_T=auto':<25} {'Auto Value':<12} {'Speedup':<10}")
    print("-" * 90)
    
    all_speedups = []
    
    for T in sequence_lengths:
        # Create test inputs
        k, u, w, g, initial_state, cu_seqlens = create_test_inputs(T, device=device)
        
        # Benchmark with chunknum_T=1
        time_1, std_1 = benchmark_kernel(k, u, w, g, initial_state, cu_seqlens, chunknum_T=1)
        
        # Benchmark with chunknum_T=None
        time_auto, std_auto = benchmark_kernel(k, u, w, g, initial_state, cu_seqlens, chunknum_T=None)
        
        # Calculate what chunknum_T was auto-selected
        B, _, Hg, K = k.shape
        V = u.shape[-1]
        H = u.shape[-2]
        N = 1
        
        estimated_bv = 32
        grid_dim_0 = -(-V // estimated_bv)
        grid_dim_1 = N * H
        base_grid_size = grid_dim_0 * grid_dim_1
        
        min_chunknum_t = max(1, -(-device_props.multi_processor_count // base_grid_size))
        chunknum_T_auto = 1
        while chunknum_T_auto < min_chunknum_t:
            chunknum_T_auto *= 2
        
        speedup = time_1 / time_auto
        all_speedups.append(speedup)
        
        print(f"{T:<12} {time_1:>8.3f} ± {std_1:>5.3f} ms   {time_auto:>8.3f} ± {std_auto:>5.3f} ms   {chunknum_T_auto:<12} {speedup:>6.2f}x")
    
    # Summary
    print("\n" + "=" * 80)
    print("Summary:")
    avg_speedup = np.mean(all_speedups)
    print(f"Average speedup across all sequence lengths: {avg_speedup:.2f}x")
    print(f"Grid base size: {grid_dim_0} * {grid_dim_1} = {base_grid_size}")
    print(f"Minimum speedup: {min(all_speedups):.2f}x")
    print(f"Maximum speedup: {max(all_speedups):.2f}x")

if __name__ == "__main__":
    main()
```

</details>

## Test Result

```
Running on: NVIDIA H200 (SM count: 132)
================================================================================
Testing different sequence lengths...
Seq Length   chunknum_T=1              chunknum_T=auto           Auto Value   Speedup   
------------------------------------------------------------------------------------------
1024            0.139 ± 0.011 ms      0.096 ± 0.001 ms   8              1.45x
2048            0.212 ± 0.002 ms      0.117 ± 0.001 ms   8              1.81x
4096            0.364 ± 0.037 ms      0.156 ± 0.001 ms   8              2.33x
8192            0.645 ± 0.004 ms      0.231 ± 0.002 ms   8              2.80x
16384           1.233 ± 0.051 ms      0.379 ± 0.002 ms   8              3.25x

================================================================================
```
Default vllm chunked prefill size corresponds to 8192 and has **2.80x speedup**